backend/requirements.txt
# ──────────────────────────────────────────────
#  AutoDarts Vision – Backend dependencies
# ──────────────────────────────────────────────

# Web framework & server
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Image handling
Pillow==11.2.1

# YOLOv8 (comment out if you don't have a GPU and want a CPU-only install)
ultralytics==8.3.106

# Data validation (bundled with FastAPI, pinned for reproducibility)
pydantic==2.11.1
